requirements.txt
pillow
jupyter
matplotlib